+++ web/htdocs/js/ui/parser/ActionProperties.js
@@ -3,29 +3,32 @@
  */
 class ActionProperties {
     
-    #actionDefinition = null;
-    #inputs = null;
+    actionDefinition = null;
+    inputs = null;
+    parserFrontend = null;
+    controller = null;
+
+    #messages = null;
+    #pagers = null;
     #oldProperties = null;
     #advancedRows = null;
-    #parserFrontend = null;
-    #messages = null;
-    #pages = null;
-    #controller = null;
-    #pageButtons = [];
 
     constructor(controller, parserFrontend, actionDefinition, oldProperties = null, messages = []) {
-        this.#controller = controller;
-        this.#parserFrontend = parserFrontend;
-        this.#actionDefinition = actionDefinition;
+        this.controller = controller;
+        this.parserFrontend = parserFrontend;
+        this.actionDefinition = actionDefinition;
+        
         this.#oldProperties = oldProperties;
         this.#messages = messages;
+        
+        this.#pagers = new PagerProperties(this)
     }
 
     /**
      * Initialize after adding to DOM
      */
     async init() {
-        if (this.#pages) await this.#pages.init();
+        await this.#pagers.init();
     }
 
     /**
@@ -33,7 +36,7 @@ class ActionProperties {
      */
     async get() {
         this.#advancedRows = [];
-        this.#inputs = new Map();
+        this.inputs = new Map();
 
         let holdInput = null;
         let assignInput = null;
@@ -45,6 +48,7 @@ class ActionProperties {
          * Take over old values from the old props object, if different from the default
          */
         async function takeOverValues(input, param) {
+            if (param.meta.type() == "text") return;
             if (!that.#oldProperties) return;
                 
             const oldParam = that.#oldProperties.getParameterDefinition(param.name);
@@ -75,7 +79,7 @@ class ActionProperties {
 
         const that = this;
         const parameters = await Promise.all(
-            this.#actionDefinition.parameters
+            this.actionDefinition.parameters
             .sort(function(a, b) {
                 return (a.meta.data.advanced ? 1 : 0) + (b.meta.data.advanced ? -1 : 0);
             })
@@ -83,7 +87,7 @@ class ActionProperties {
                 async (param) => {
                     const input = await this.#createInput(param);
 
-                    that.#inputs.set(param.name, input);
+                    that.inputs.set(param.name, input);
 
                     // Take over old values from the old props object, if different from the default
                     await takeOverValues(input, param);
@@ -148,12 +152,10 @@ class ActionProperties {
         
         let tbody = null;
 
-        const pagerActions = await this.#parserFrontend.parser.pagerActions();
-        
         const ret = $('<div class="action-properties" />').append(
             // Comment
             $('<div class="action-header" />')
-            .text(this.#actionDefinition.meta.getDisplayName()),
+            .text(this.actionDefinition.meta.getDisplayName()),
             
             $('<div class="action-comment" />')
             .html(this.#getActionComment()),
@@ -167,7 +169,7 @@ class ActionProperties {
                     tbody = $('<tbody />').append(
 
                         // Hold option
-                        (this.#actionDefinition.meta.data.target != "AdafruitSwitch") ? null :
+                        (this.actionDefinition.meta.data.target != "AdafruitSwitch") ? null :
                         withComment(
                             $('<tr />').append(                            
                                 $('<td />').append(
@@ -209,7 +211,7 @@ class ActionProperties {
         
                                 // Input
                                 $('<td />').append(
-                                    pagerProxyInput = (await this.#createPageProxyInput())
+                                    pagerProxyInput = (await this.#pagers.createProxyInput())
                                 ),
                             )
                             .hide(),
@@ -224,17 +226,10 @@ class ActionProperties {
             ),
 
             // Pager buttons
-            ...(
-                !pagerActions.length ? [] : [
-                    $('<div class="action-header" />')
-                    .text("Assign to Page:"),
-                    
-                    await this.#getPagerButtons(pagerActions)
-                ]
-            )
+            ...(await this.#pagers.getButtons())
         );
 
-        if (this.#actionDefinition.name != "PagerAction") {
+        if (this.actionDefinition.name != "PagerAction") {
             assignRow.hide();
 
             this.#advancedRows.push({
@@ -243,9 +238,9 @@ class ActionProperties {
             });
         }
 
-        if (this.#actionDefinition.name == "PagerAction.proxy") {
+        if (this.actionDefinition.name == "PagerAction.proxy") {
             pagerProxyRow.show();
-            this.#inputs.set("pager", pagerProxyInput);
+            this.inputs.set("pager", pagerProxyInput);
         }
 
         // Advanced parameters: Show all button
@@ -264,7 +259,7 @@ class ActionProperties {
 
                                 advRow.hide();
                             } catch (e) {
-                                that.#controller.handle(e);
+                                that.controller.handle(e);
                             }
                         })
                     )
@@ -273,15 +268,15 @@ class ActionProperties {
         }
 
         // Hold input
-        if (this.#actionDefinition.meta.data.target == "AdafruitSwitch") {
-            this.#inputs.set("hold", holdInput);
+        if (this.actionDefinition.meta.data.target == "AdafruitSwitch") {
+            this.inputs.set("hold", holdInput);
             if (this.#oldProperties) {
                 this.setHold(this.#oldProperties.hold());            
             }
         }
 
         // Assign input
-        this.#inputs.set("assign", assignInput);        
+        this.inputs.set("assign", assignInput);        
         // if (this.#oldProperties) {
         //     this.setAssign(this.#oldProperties.assign());            
         // }
@@ -290,263 +285,13 @@ class ActionProperties {
     }
 
     /**
-     * Creates the pager proxy input for PagerAction.proxy
-     */
-    async #createPageProxyInput() {
-        const pagerActions = await this.#parserFrontend.parser.pagerActions();
-        const that = this;
-
-        return $('<select />').append(
-            pagerActions.map((option) => {
-                return $('<option value="' + option.assign + '" />')
-                .text(option.assign)
-            })                        
-        )
-        .on('change', async function() {
-            await that.#update();
-        })
-    }
-
-    /**
      * Returns the default assign value
      */
     async #getDefaultAssign() {
-        if (this.#actionDefinition.name == "PagerAction") {
-            return this.#getNextPagerAssign();
+        if (this.actionDefinition.name == "PagerAction") {
+            return this.#pagers.getNextPagerAssign();
         }
         return "";
-    }
-
-    /**
-     * Returns an unused pager assign target name
-     */
-    async #getNextPagerAssign() {
-        const pagerActions = await this.#parserFrontend.parser.pagerActions();
-
-        function pagerExists(assign) {
-            for (const pager of pagerActions) {
-                if (pager.assign == assign) return true;
-            }
-            return false;
-        }
-
-        let ret = "_pager";
-        let cnt = 2;
-        while (pagerExists(ret)) {
-            ret = "_pager" + cnt++;
-        }
-        return ret;
-    }
-
-    /**
-     * Strip the page text from quotes
-     */
-    #stripPageText(page) {
-        return page.text ? page.text.replaceAll('"', "").replaceAll("'", "") : null;
-    }
-
-    /**
-     * Returns DOM for the pager buttons
-     */
-    async #getPagerButtons(pagerActions) {
-        this.#pageButtons = [];
-        const that = this;
-
-        function getPageText(actionCallProxy, el) {
-            const pageText = that.#stripPageText(el) || el.id;
-
-            if (pagerActions.length == 1) {
-                return "" + pageText;
-            }
-            
-            return (actionCallProxy.assign ? (actionCallProxy.assign + "|") : "") + pageText;
-        }
-
-        const noPageButton = $('<span class="button"/>')
-            .text("No Page")
-            .on('click', async function() {
-                try {
-                    await that.#setPageParameters(null, null);
-
-                } catch (e) {
-                    that.#controller.handle(e);
-                }
-            });
-        
-        this.#pageButtons.push({
-            pager: "None",
-            page: "None",
-            button: noPageButton
-        })
-
-        return $('<div class="action-pages" />').append(
-            // $('<div class="action-pages-comment" />')
-            // .text('To assign this action to a page, use these buttons:'),
-
-            noPageButton,
-
-            (
-                await Promise.all(
-                    pagerActions
-                    .map(
-                        async (item) => {
-                            const pagesArg = JSON.parse(item.arguments()).filter((e) => e.name == "pages");
-                            if (!pagesArg || !pagesArg.length) return null;
-
-                            const pages = pagesArg[0].value;
-                            return (
-                                await Promise.all(
-                                    pages.map(async (el) => 
-                                        {
-                                            let pageColorIcon = null;
-                                            const button = $('<span class="button"/>').append(
-                                                // Page color icon
-                                                pageColorIcon = $('<span class="page-icon" />'),
-
-                                                // Page ID
-                                                $('<span />')
-                                                .text(getPageText(item, el))
-                                            )
-                                            .on('click', async function() {
-                                                try {
-                                                    await that.#setPageParameters(item, el);
-
-                                                } catch (e) {
-                                                    that.#controller.handle(e);
-                                                }
-                                            });
-
-                                            const color = await that.#parserFrontend.icons.getPageColor(item.assign, el.id);
-                                            if (color) {
-                                                pageColorIcon.css('background-color', "rgb" + color);
-                                            } else {
-                                                pageColorIcon.hide()
-                                            }
-
-                                            that.#pageButtons.push({
-                                                pager: item.assign,
-                                                page: el.id,
-                                                button: button
-                                            })                    
-
-                                            return button;
-                                        }
-                                    )
-                                )
-                            )
-                        }
-                    )
-                )
-            ).flat()
-        );
-    }
-
-    /**
-     * Update the state of the pager buttons
-     */
-    async #updatePagerButtons() {
-        const pager = this.#getPager();
-        const page = this.#getPage();
-        
-        for (const button of this.#pageButtons) {
-            const selected = (button.pager == pager && button.page == page);
-            button.button.toggleClass("page-selected", selected);
-        }
-    }
-
-    /**
-     * Returns the current pager (derived from the enable callback)
-     */
-    #getPager() {
-        if (!this.#inputs.has("enable_callback")) return null;
-        return this.#extractPager(this.#inputs.get("enable_callback").val());
-    }
-
-    /**
-     * Extracts the pager name from _pager.xxx
-     */
-    #extractPager(name) {
-        const splt = name.split(".");
-        if (!splt.length == 2) return null;
-
-        return splt[0];
-    }
-
-    /**
-     * Returns the currently set page (parameter "id")
-     */
-    #getPage() {
-        if (!this.#inputs.has("id")) return null;
-        return this.#inputs.get("id").val();
-    }
-
-    /**
-     * Sets the parameters for the given page of the given PagerAction.
-     */
-    async #setPageParameters(actionCallProxy, page) {
-        await this.setArgument("enable_callback", ((actionCallProxy && actionCallProxy.assign) ? (actionCallProxy.assign + ".enable_callback") : "None"));
-        await this.setArgument("id", (page && page.id) ? page.id : "None");
-
-        // this.#showParameter("enable_callback");
-        // this.#showParameter("id");
-    }
-
-    /**
-     * Updates the page inputs to the currently available pages
-     */
-    async #updatePagesInputs() {
-        const that = this;
-
-        /**
-         * Add pages as options to a select input
-         */
-        function addPagesToInput(input, pages) {
-            input.empty().append(                
-                pages.map((item) => {
-                    const pageText = that.#stripPageText(item);
-                    return $('<option value="' + item.id + '" />')
-                    .text(pageText ? (item.id + " " + pageText) : item.id)
-                })
-            );
-        }
-
-        // Pager
-        if (this.#pages && this.#actionDefinition.name == "PagerAction") {
-            const pages = this.#pages.get();
-            pages.push({
-                id: "None"                
-            })
-
-            const input = this.#inputs.get("select_page");
-            if (input) {
-                const value = input.val() || "None";
-                addPagesToInput(input, pages);
-                input.val(value);
-            }
-        }
-
-        // Pager proxy
-        if (this.#actionDefinition.name == "PagerAction.proxy") {
-            const pagerProxy = this.#inputs.get("pager").val();
-            
-            if (pagerProxy) {
-                const pager = await this.#parserFrontend.parser.getPagerAction(pagerProxy)
-                if (pager) {
-                    const pages = pager.argument("pages");
-                    if (pages) {
-                        const input = this.#inputs.get("page_id");
-                        if (input) {
-                            const value = input.val();
-                            addPagesToInput(
-                                input, 
-                                pages.map((item) => item.toJs())
-                            );
-                            input.val(value);
-                        }
-                    }
-                }
-            }
-        }
     }
 
     /**
@@ -556,22 +301,22 @@ class ActionProperties {
         const that = this;
 
         function getName() {
-            if (that.#actionDefinition.name == "PagerAction.proxy") {
-                const pagerProxy = that.#inputs.get("pager").val();
+            if (that.actionDefinition.name == "PagerAction.proxy") {
+                const pagerProxy = that.pagerProxy();
                 
                 if (pagerProxy) {
-                    return that.#actionDefinition.name.replace("PagerAction", pagerProxy)
+                    return that.actionDefinition.name.replace("PagerAction", pagerProxy)
                 }
             }
-            return that.#actionDefinition.name;
+            return that.actionDefinition.name;
         }
 
         return {
             name: getName(),
-            assign: this.#inputs.get('assign').val(),
-            arguments: this.#actionDefinition.parameters
+            assign: this.inputs.get('assign').val(),
+            arguments: this.actionDefinition.parameters
                 .filter((param) => {
-                    const input = that.#inputs.get(param.name);
+                    const input = that.inputs.get(param.name);
                     if (!input) throw new Error("No input for param " + param.name + " found");
         
                     const value = that.#getInputValue(input, param);
@@ -579,7 +324,7 @@ class ActionProperties {
                     return !param.hasOwnProperty("default") || (value != param.default);
                 })
                 .map((param) => {
-                    const input = that.#inputs.get(param.name);
+                    const input = that.inputs.get(param.name);
                     if (!input) throw new Error("No input for param " + param.name + " found");
         
                     return {
@@ -594,8 +339,8 @@ class ActionProperties {
      * Returns if the user selected hold or not (JS bool, no python value)
      */
     hold() {
-        if (this.#actionDefinition.meta.data.target == "AdafruitSwitch") {
-            return !!this.#inputs.get("hold").prop('checked');
+        if (this.actionDefinition.meta.data.target == "AdafruitSwitch") {
+            return !!this.inputs.get("hold").prop('checked');
         }
         return false;
     }
@@ -604,45 +349,47 @@ class ActionProperties {
      * Sets the hold input
      */
     setHold(hold) {
-        if (this.#actionDefinition.meta.data.target != "AdafruitSwitch") {
+        if (this.actionDefinition.meta.data.target != "AdafruitSwitch") {
             return;
         }
-        this.#inputs.get("hold").prop('checked', !!hold)
+        this.inputs.get("hold").prop('checked', !!hold)
     }
 
     /**
      * Returns the assign value if set
      */
     assign() {
-        return this.#inputs.get("assign").val();
+        if (!this.inputs.has("assign")) return null;
+        return this.inputs.get("assign").val();
     }
 
     /**
      * Sets the assign input
      */
     setAssign(assign) {
-        this.#inputs.get("assign").val(assign);
+        this.inputs.get("assign").val(assign);
     }
 
     /**
      * Returns the pager proxy value if set
      */
     pagerProxy() {
-        return this.#inputs.get("pager").val();
+        if (!this.inputs.has("pager")) return null;
+        return this.inputs.get("pager").val();
     }
 
     /**
      * Sets the pager proxy input
      */
     setPagerProxy(proxy) {
-        this.#inputs.get("pager").val(proxy);
+        this.inputs.get("pager").val(proxy);
     }
 
     /**
      * Sets the input values to the passed arguments list's values
      */
     async setArguments(args) {
-        await this.#update();
+        await this.update();
 
         for (const arg of args) {
             await this.setArgument(arg.name, arg.value);
@@ -655,25 +402,25 @@ class ActionProperties {
             }
         }
 
-        await this.#update();
+        await this.update();
     }
 
     /**
      * Set the value of a parameter input
      */
     async setArgument(name, value) {
-        await this.#update();
+        await this.update();
 
         // Get parameter definition first
         const param = this.getParameterDefinition(name);
         if (!param) throw new Error("Parameter " + name + " not found");
 
-        const input = this.#inputs.get(param.name);
+        const input = this.inputs.get(param.name);
         if (!input) throw new Error("No input for param " + param.name + " found");
 
         await this.#setInputValue(input, param, value);
 
-        await this.#update();
+        await this.update();
     }
 
     /**
@@ -691,7 +438,7 @@ class ActionProperties {
      * Searches a parameter mode by name
      */
     getParameterDefinition(name) {
-        for (const param of this.#actionDefinition.parameters) {
+        for (const param of this.actionDefinition.parameters) {
             if (param.name == name) return param;
         }
         return null;
@@ -701,8 +448,8 @@ class ActionProperties {
      * Determine the comment for the action
      */
     #getActionComment() {
-        if (!this.#actionDefinition.comment) return "No information available";
-        let comment = "" + this.#actionDefinition.comment;
+        if (!this.actionDefinition.comment) return "No information available";
+        let comment = "" + this.actionDefinition.comment;
 
         if (comment.slice(-1) != ".") comment += ".";
 
@@ -721,20 +468,19 @@ class ActionProperties {
     /**
      * Update the UI
      */
-    async #update() {
-        await this.#updatePagerButtons();
-        await this.#updatePagesInputs();
+    async update() {
+        await this.#pagers.update();
     }
 
     /**
      * Generates the DOM for one parameter
      */
     async #createInput(param) {
-        const type = this.#deriveType(param);
+        const type = param.meta.type();
 
         const that = this;
         async function onChange() {            
-            await that.#update();
+            await that.update();
         }
 
         switch(type) {
@@ -772,8 +518,7 @@ class ActionProperties {
                 
             case 'pages': {
                 // Dedicated type for the pager actions's "pages" parameter
-                this.#pages = new PagesList(this.#controller, onChange);
-                return this.#pages.create()
+                return this.#pagers.getPagesList(onChange);
             }
         }        
 
@@ -788,13 +533,13 @@ class ActionProperties {
      */
     async #createAdditionalColorInputOptions(input, param) {
         const that = this;
-        const type = this.#deriveType(param);
+        const type = param.meta.type();
         if (type != "color") return [];
         
         let colorInput = null;
 
         async function updateColorInput() {
-            const color = await that.#parserFrontend.parser.resolveColor(input.val());
+            const color = await that.parserFrontend.parser.resolveColor(input.val());
             if (color) {
                 colorInput.val(Tools.rgbToHex(color))
             }
@@ -802,7 +547,7 @@ class ActionProperties {
 
         const ret = [
             $('<select class="parameter-option" />').append(
-                (await this.#parserFrontend.parser.getAvailableColors())
+                (await this.parserFrontend.parser.getAvailableColors())
                 .concat([{
                     name: "Select color..."
                 }])
@@ -843,7 +588,7 @@ class ActionProperties {
         const param = this.getParameterDefinition(name);
         if (!param) return null;
 
-        const input = this.#inputs.get(param.name);
+        const input = this.inputs.get(param.name);
         if (!input) return null;
 
         return this.#getInputValue(input, param);        
@@ -853,32 +598,24 @@ class ActionProperties {
      * Converts the input values to action argument values
      */
     #getInputValue(input, param) {
-        const type = this.#deriveType(param);
+        const type = param.meta.type();
 
         switch(type) {
             case "bool": return input.prop('checked') ? "True" : "False";
-            case "pages": return this.#pages.get();
+            case "pages": return this.#pagers.pages.get();
         }        
 
         let value = input.val();
         if (value == "") value = param.meta.getDefaultValue();
 
-        function getTextTypeValue(v) {
-            // TODO Allow other values, move to ParameterMeta
-            return Tools.autoQuote(v);
-        }
-
-        switch (type) {
-            case "text": return getTextTypeValue(value)
-        }
-        return value;
+        return param.meta.convertInput(value);
     }
 
     /**
      * Sets the input value according to an argumen/parameter value
      */
     async #setInputValue(input, param, value) {
-        const type = this.#deriveType(param);
+        const type = param.meta.type();
 
         switch(type) {
             case "bool": 
@@ -887,7 +624,7 @@ class ActionProperties {
                 break;
 
             case "pages":
-                await this.#pages.set(value)
+                await this.#pagers.pages.set(value)
                 break;
 
             default:
@@ -911,22 +648,5 @@ class ActionProperties {
                 .text(option.name)
             )
         ) 
-    }
-
-    /**
-     * Tries to derive the parameter type from its default value. Returns null if not successful.
-     */
-    #deriveType(param) {
-        if (param.meta.data.type) return param.meta.data.type;;
-
-        const defaultValue = param.meta.getDefaultValue();
-        switch (defaultValue) {
-            case "False": return "bool";
-            case "True": return "bool";            
-        }
-
-        if (parseInt(defaultValue)) return "int";
-
-        return null;
     }
 }