+++ CHANGELOG.md
@@ -3,10 +3,10 @@
 - Added option to directly select pages (see proxy method of PagerAction, demonstrated in some new examples for the MIDI Captain 10)
 
 ### Emulator 2.3.5.5
-- Auto-add quotes to page text fields
 - Added color input help
 - Added explanation to the client select dialog
 - Auto-add quotes to text parameters and page texts
+- Prevent assigning pagers to their own pages
 
 ### Emulator 2.3.5.4
 - Added full support for all Paging functionality (see Pager and Select Page actions)